fix: resolve #90 — SSRF guard aplicado em createTransport e connectWithFallback do MCPAdapter
